Refactor corporate dashboard into modular components
- Replaced the existing CorporateDashboard component with a new modular implementation.
- Created separate components for metrics, tabs, and individual tab content (Overview, Performance, Risk, Reports).
- Maintained backward compatibility by re-exporting the new implementation from the original path.
- Updated sidebar and header integration within the new structure.
- Improved code organization and readability by separating concerns into distinct files.
